fleet template: add || true to some things that could normally fail
Running Docker with --rm and an ExecStop docker rm will cause false
failures of images. This has caused problems when I have been deploying
a test application to my instance of Panamax. This fixes the problem by
adding `|| true` to commands that *could* normally return a non-0 exit
code so that the web UI and other parts of Panamax do not get confused.