Add deployment target metadata refresh endpoint.
So there's a lot going on here. Agents can return their metadata, which
includes the metadata from their connected adapter. This involves a lot
of requests when you go from the UI all the way to the adapter, so we
want to persist it. There are two models that can be easily confused but
are distinct. RemoteAgentMetadata is an ApiModel model that holds the
values of the agent's API response, and DeploymentTargetMetadata is the
persistence in the API.