Nest metadata within metadata refresh response.
This accomplishes two things: it makes room for success/error
information in our response, and makes this endpoint easier to use in
ActiveResource. It's going to be hard to 'create' a refresh but return a
metadata object. If the refresh object that is returned contains
metadata, though, that jives better with ActiveResource.