Explicitly define required fixtures in tests.
The global fixture setting causes dozens of rows to be created and then
rolled back on every test. I feel like it's clearer when you're testing
(or reading the test) about what's involved with a system when you have
to state which fixtures you need in the context it will be used. It has
a side-effect of being significantly faster.