Our Circle builds include the following warning:
Your version of Simplecov has a serious bug affecting exit codes! It can
cause your build to pass even when you have failing tests

There is a slightly newer version of coveralls (which is what brings in
simplecov), but it addresses a couple of edge cases that we don't seem
to care about. It works fine with the newer version of simplecov, so in
the interest of changing the fewest things I'm just upgrading simplecov.